feat(core): Use package snapshots
These changes allow `straight` to download a tarball archive instead of
a full clone of the git repository.

* lisp/core/cli/packages.el (zenit--cli-recipes-update): Ensure, that
recipe local repos are actual git repos.
(zenit-cli-packages-update): Correctly handle snapshots during updating.

* lisp/core/lib/zenit-lib-packages.el (zenit/bump-package): If a package
is intalled as a snapshot, the local-repo does not contain the git log.
Thus we check for the `.straight-commit` file and if found, we do a
shallow clone of the repo and only fetch the history to populate the
commit candidate list.

* lisp/core/zenit-packages.el: Set
`straight-vc-use-snapshot-installation` to non-nil if `tar` is installed
on the system.

* straight/versions/core.el: Pin `straight` to the latest version.