fix(:core package): Better handling of local lockfile versions
* lisp/core/zenit-packages.el (+straight--read-local-lockfile-a):
(+straight--vc-clone-prefer-local-profile-a): Bind
`straight-current-profile` during cloning to (one of) the declared
profiles and check if the package was declared as a local package. If
so, prefer the local lockfile, even if absent and thus, unpinning the
package.
Furthermore, always push `straight-current-profile` into the cache.
Otherwise, dep packages get not registered under another profile (if
any).